+++ compiler-core/src/format.rs
@@ -19,6 +19,7 @@ use smol_str::SmolStr;
 use std::sync::Arc;
 use vec1::Vec1;
 
+use crate::type_::Deprecation;
 use camino::Utf8Path;
 
 const INDENT: isize = 2;
@@ -212,8 +213,9 @@ impl<'comments> Formatter<'comments> {
                 parameters: args,
                 type_ast: resolved_type,
                 public,
+                deprecation,
                 ..
-            }) => self.type_alias(*public, alias, args, resolved_type),
+            }) => self.type_alias(*public, alias, args, resolved_type, deprecation),
 
             Definition::CustomType(ct) => self.custom_type(ct),
 
@@ -459,8 +461,12 @@ impl<'comments> Formatter<'comments> {
         name: &'a str,
         args: &'a [SmolStr],
         typ: &'a TypeAst,
+        deprecation: &'a Deprecation,
     ) -> Document<'a> {
-        let head = pub_(public).append("type ").append(name);
+        // @deprecated attribute
+        let head = self.deprecation_attr(deprecation);
+
+        let head = head.append(pub_(public)).append("type ").append(name);
 
         let head = if args.is_empty() {
             head
@@ -470,6 +476,18 @@ impl<'comments> Formatter<'comments> {
 
         head.append(" =")
             .append(line().append(self.type_ast(typ)).group().nest(INDENT))
+    }
+
+    fn deprecation_attr<'a>(&mut self, deprecation: &'a Deprecation) -> Document<'a> {
+        match deprecation {
+            Deprecation::NotDeprecated => "".to_doc(),
+            Deprecation::Deprecated { message } => ""
+                .to_doc()
+                .append("@deprecated(\"")
+                .append(message)
+                .append("\")")
+                .append(line()),
+        }
     }
 
     fn fn_arg<'a, A>(&mut self, arg: &'a Arg<A>) -> Document<'a> {
@@ -483,17 +501,8 @@ impl<'comments> Formatter<'comments> {
     }
 
     fn statement_fn<'a>(&mut self, function: &'a Function<(), UntypedExpr>) -> Document<'a> {
-        let attributes = docvec![];
-
         // @deprecated attribute
-        let attributes = match &function.deprecation {
-            type_::Deprecation::NotDeprecated => attributes,
-            type_::Deprecation::Deprecated { message } => attributes
-                .append("@deprecated(\"")
-                .append(message)
-                .append("\")")
-                .append(line()),
-        };
+        let attributes = self.deprecation_attr(&function.deprecation);
 
         // @external attribute
         let external = |t: &'static str, m: &'a str, f: &'a str| {
@@ -1078,7 +1087,12 @@ impl<'comments> Formatter<'comments> {
 
     pub fn custom_type<'a, A>(&mut self, ct: &'a CustomType<A>) -> Document<'a> {
         let _ = self.pop_empty_lines(ct.location.end);
-        let doc = pub_(ct.public)
+
+        // @deprecated attribute
+        let doc = self.deprecation_attr(&ct.deprecation);
+
+        let doc = doc
+            .append(pub_(ct.public))
             .to_doc()
             .append(if ct.opaque { "opaque type " } else { "type " })
             .append(if ct.parameters.is_empty() {