+++ compiler-core/src/type_/tests/warnings.rs
@@ -915,6 +915,115 @@ pub fn b() {
 }
 
 #[test]
+fn deprecated_imported_function() {
+    assert_warning!(
+        (
+            "package",
+            "module",
+            r#"@deprecated("Don't use this!") pub fn a() { Nil }"#
+        ),
+        r#"
+import module
+
+pub fn a() {
+  module.a
+}
+"#
+    );
+}
+
+#[test]
+fn deprecated_imported_call_function() {
+    assert_warning!(
+        (
+            "package",
+            "module",
+            r#"@deprecated("Don't use this!") pub fn a() { Nil }"#
+        ),
+        r#"
+import module
+
+pub fn a() {
+  module.a()
+}
+"#
+    );
+}
+
+#[test]
+fn deprecated_imported_unqualified_function() {
+    assert_warning!(
+        (
+            "package",
+            "module",
+            r#"@deprecated("Don't use this!") pub fn a() { Nil }"#
+        ),
+        r#"
+import module.{a}
+
+pub fn b() {
+  a
+}
+"#
+    );
+}
+
+#[test]
+fn deprecated_type_used_in_alias() {
+    assert_warning!(
+        r#"
+@deprecated("Don't use this!")
+pub type Cat {
+    Cat(name: String)
+}
+
+pub type Dog = Cat
+        "#
+    );
+}
+
+#[test]
+fn deprecated_type_used_as_arg() {
+    assert_warning!(
+        r#"
+@deprecated("Don't use this!")
+pub type Cat {
+    Cat(name: String)
+}
+
+pub fn cat_name(cat: Cat) {
+  cat.name
+}
+        "#
+    );
+}
+
+#[test]
+fn deprecated_type_used_as_case_clause() {
+    assert_warning!(
+        r#"
+@deprecated("The type Animal has been deprecated.")
+pub type Animal {
+    Cat
+    Dog
+}
+
+pub fn sound(animal) -> String {
+  case animal {
+    Dog -> "Woof"
+    Cat -> "Meow"
+  }
+}
+
+pub fn main(){
+    let cat = Cat
+    sound(cat)
+}
+        "#
+    );
+}
+
+#[test]
 fn deprecated_bit_array_type() {
     assert_warning!(r#"pub type B = BitString"#);
 }